Reorganize repo in preparation for TH-based instance derivation
