+++ src/Test/Goblin/TH.hs
@@ -1,0 +1,3 @@
+{-# LANGUAGE TemplateHaskell       #-}
+module Test.Goblin.TH where
+
